chore(discovery): clean up and address reviews
