+++ src/main/resources/application-dev.properties
@@ -19,9 +19,6 @@ cryostat.discovery.podman.enabled=true
 cryostat.discovery.docker.enabled=true
 cryostat.discovery.kubernetes.enabled=true
 
-cryostat.discovery.k8s.namespaces=
-kubernetes.service.host=
-
 quarkus.datasource.devservices.enabled=true
 quarkus.datasource.devservices.image-name=quay.io/cryostat/cryostat-db
