+++ src/main/resources/application-test.properties
@@ -4,10 +4,6 @@ cryostat.discovery.jdp.enabled=true
 cryostat.discovery.podman.enabled=true
 cryostat.discovery.docker.enabled=true
 cryostat.discovery.kubernetes.enabled=true
-cryostat.discovery.k8s.port-names=
-cryostat.discovery.k8s.port-numbers=
-cryostat.discovery.k8s.namespaces=
-kubernetes.service.host=
 
 quarkus.test.env.JAVA_OPTS_APPEND=-Dquarkus.http.host=0.0.0.0 -Djava.util.logging.manager=org.jboss.logmanager.LogManager -Dcom.sun.management.jmxremote -Dcom.sun.management.jmxremote.port=9091 -Dcom.sun.management.jmxremote.rmi.port=9091 -Djava.rmi.server.hostname=127.0.0.1 -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.ssl=false -Dcom.sun.management.jmxremote.local.only=false
