+++ src/main/resources/application.properties
@@ -11,9 +11,11 @@ cryostat.discovery.plugins.jwt.signature.algorithm=HS256
 cryostat.discovery.plugins.jwt.encryption.algorithm=dir
 cryostat.discovery.plugins.jwt.encryption.method=A256GCM
 cryostat.discovery.kubernetes.enabled=false
-cryostat.discovery.k8s.port-names=
-cryostat.discovery.k8s.port-numbers=
-cryostat.discovery.k8s.namespaces=
+cryostat.discovery.kubernetes.port-names=
+cryostat.discovery.kubernetes.port-numbers=
+cryostat.discovery.kubernetes.namespaces=
+cryostat.discovery.kubernetes.namespace-path=/var/run/secrets/kubernetes.io/serviceaccount/namespace
+cryostat.discovery.kubernetes.resync-period=30s
 kubernetes.service.host=
 
 quarkus.test.integration-test-profile=test
